# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks for the current user
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          description: Not signed in
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new plant care task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid input
        "401":
          description: Not signed in
  /tasks/{id}/complete:
    patch:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as completed (resets the reminder cycle)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Not signed in
        "404":
          description: Task not found
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update an existing task's fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error
        "404":
          description: Task not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Permanently delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not signed in
        "404":
          description: Task not found
...
  /users:
    get:
      tags: [users]
      operationId: listUsers
      summary: List all accounts that have ever signed in (owner only)
      responses:
        "200":
          description: All accounts, owner first, then by sign-up date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedUser"
        "401":
          description: Not signed in
        "403":
          description: Signed in but not the owner
  /users/{id}:
    patch:
      tags: [users]
      operationId: updateUserApproval
      summary: Approve or revoke an account (owner only)
      description: >
        Grants or revokes app access for an account. Revoking also deletes
        all sessions of that account so it is signed out everywhere. The
        owner account cannot be modified.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserApprovalInput"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "400":
          description: Invalid body or attempt to modify the owner account
        "401":
          description: Not signed in
        "403":
          description: Signed in but not the owner
        "404":
          description: Account not found
    delete:
      tags: [users]
      operationId: deleteUser
      summary: Remove an account and all its sessions (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
        "400":
          description: Attempt to remove the owner account
        "401":
          description: Not signed in
        "403":
          description: Signed in but not the owner
        "404":
          description: Account not found
...
components:
...
  schemas:
...
    ManagedUser:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - profileImageUrl
        - approved
        - isOwner
        - createdAt
        - username
        - leavesCount
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        approved:
          type: boolean
        isOwner:
          type: boolean
        createdAt:
          type: string
          format: date-time
        username:
          type: [string, 'null']
        leavesCount:
          type: integer
    TaskActionType:
      type: string
      description: The type of care action to perform.
      enum:
        - Gießen
        - Besprühen
        - Düngen
        - Pflanze drehen
    TaskIntervalUnit:
      type: string
      description: The unit of the reminder interval.
      enum:
        - Tage
        - Wochen
        - Monate
    TaskFertilizerType:
      type: string
      description: The type of fertilizer (only relevant when actionType is "Düngen").
      enum:
        - Biologischer Dünger
        - Mineralischer Dünger
        - Manuell
    Task:
      type: object
      required:
        - id
        - userId
        - plantId
        - actionType
        - intervalValue
        - intervalUnit
        - reminderTime
        - isActive
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: string
        plantId:
          type: integer
        actionType:
          $ref: "#/components/schemas/TaskActionType"
        intervalValue:
          type: integer
          minimum: 1
          maximum: 60
        intervalUnit:
          $ref: "#/components/schemas/TaskIntervalUnit"
        reminderTime:
          type: string
          description: Local time for reminder in HH:MM format (e.g. "08:00").
        isActive:
          type: boolean
        lastCompletedAt:
          type: string
          nullable: true
          description: ISO timestamp of the last time this task was marked as done.
        fertilizerType:
          nullable: true
          description: Fertilizer sub-type, only set when actionType is "Düngen".
          oneOf:
            - $ref: "#/components/schemas/TaskFertilizerType"
        fertilizerCustomName:
          type: string
          nullable: true
          description: Custom fertilizer name when fertilizerType is "Manuell".
        createdAt:
          type: string
    UpdateTaskInput:
      type: object
      description: All fields optional – only provided fields are updated.
      properties:
        actionType:
          $ref: "#/components/schemas/TaskActionType"
        intervalValue:
          type: integer
          minimum: 1
          maximum: 60
        intervalUnit:
          $ref: "#/components/schemas/TaskIntervalUnit"
        reminderTime:
          type: string
          description: Local time for reminder in HH:MM format (e.g. "08:00").
        fertilizerType:
          nullable: true
          description: Fertilizer sub-type, only relevant when actionType is "Düngen".
          oneOf:
            - $ref: "#/components/schemas/TaskFertilizerType"
        fertilizerCustomName:
          type: string
          nullable: true
          description: Custom fertilizer name when fertilizerType is "Manuell".
    CreateTaskInput:
      type: object
      required:
        - plantId
        - actionType
        - intervalValue
        - intervalUnit
        - reminderTime
      properties:
        plantId:
          type: integer
        actionType:
          $ref: "#/components/schemas/TaskActionType"
        intervalValue:
          type: integer
          minimum: 1
          maximum: 60
        intervalUnit:
          $ref: "#/components/schemas/TaskIntervalUnit"
        reminderTime:
          type: string
          description: Local time for reminder in HH:MM format (e.g. "08:00").
        fertilizerType:
          nullable: true
          description: Fertilizer sub-type, only set when actionType is "Düngen".
          oneOf:
            - $ref: "#/components/schemas/TaskFertilizerType"
        fertilizerCustomName:
          type: string
          nullable: true
          description: Custom fertilizer name when fertilizerType is "Manuell".
...
    UpdateUserApprovalInput:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
          description: True grants access, false revokes it (the owner account cannot be modified).
...